server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.9.24; python_full_version >= '3.6.0'
charset-normalizer==2.1.1; python_full_version >= '3.6.0'
click==8.1.3; python_version >= '3.7'
flask==2.2.2
google-api-core==2.10.1; python_version >= '3.7'
google-auth==2.12.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-appengine-logging==1.1.5; python_version >= '3.7'
google-cloud-audit-log==0.2.4; python_version >= '3.7'
google-cloud-core==2.3.2; python_version >= '3.7'
google-cloud-logging==3.2.4
google-cloud-storage==2.5.0
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.4.0; python_version >= '3.7'
googleapis-common-protos==1.56.4; python_version >= '3.7'
grpc-google-iam-v1==0.12.4; python_full_version >= '3.6.0'
grpcio-status==1.49.1
grpcio==1.49.1; python_full_version >= '3.6.0'
idna==3.4; python_version >= '3.5'
importlib-metadata==5.0.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
pendulum==2.1.2
proto-plus==1.22.1; python_full_version >= '3.6.0'
protobuf==4.21.7; python_version >= '3.7'
pyasn1-modules==0.2.8
pyasn1==0.4.8
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
pytzdata==2020.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.1
rsa==4.9; python_full_version >= '3.6.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.8.1; python_version >= '3.7'
